third_party/blink/web_tests/platform/win/external/wpt/pointerevents/compat/pointerevent_touch_target_after_pointerdown_target_removed.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] After a pointerdown listener removes its target, touchmove event should be fired on the pointerdown target
  assert_equals: expected "[ { type: pointerdown, target: Element node <div id=\\"pointerDownTarget\\"></div> }, { type: touchstart, target: Element node <div id=\\"pointerDownTarget\\"></div> }, { type: pointermove, target: Element node <div id=\\"pointerDownTargetParent\\"></div> }, { type: touchmove, target: Element node <div id=\\"pointerDownTarget\\"></div> }, { type: pointerup, target: Element node <div id=\\"pointerDownTargetParent\\"></div> }, { type: touchend, target: Element node <div id=\\"pointerDownTarget\\"></div> } ]" but got "[ { type: pointerdown, target: Element node <div id=\\"pointerDownTarget\\"></div> }, { type: touchstart, target: Element node <div id=\\"pointerDownTarget\\"></div> }, { type: pointermove, target: Element node <div id=\\"pointerDownTargetParent\\"></div> }, { type: pointerup, target: Element node <div id=\\"pointerDownTargetParent\\"></div> }, { type: touchend, target: Element node <div id=\\"pointerDownTarget\\"></div> } ]"
[FAIL] After a pointerdown listener moves the target to different position, touchmove event should be fired on the pointerdown target parent
  assert_equals: expected "[ { type: pointerdown, target: Element node <div id=\\"pointerDownTarget\\"></div> }, { type: touchstart, target: Element node <div id=\\"pointerDownTarget\\"></div> }, { type: pointermove, target: Element node <div id=\\"pointerDownTargetParent\\"></div> }, { type: touchmove, target: Element node <div id=\\"pointerDownTarget\\"></div> }, { type: pointerup, target: Element node <div id=\\"pointerDownTargetParent\\"></div> }, { type: touchend, target: Element node <div id=\\"pointerDownTarget\\"></div> } ]" but got "[ { type: pointerdown, target: Element node <div id=\\"pointerDownTarget\\"></div> }, { type: touchstart, target: Element node <div id=\\"pointerDownTarget\\"></div> }, { type: pointermove, target: Element node <div id=\\"pointerDownTargetParent\\"></div> }, { type: pointerup, target: Element node <div id=\\"pointerDownTargetParent\\"></div> }, { type: touchend, target: Element node <div id=\\"pointerDownTarget\\"></div> } ]"
[FAIL] After a touchstart listener removes its target, touchmove event should be fired on the pointerdown target
  assert_equals: expected "[ { type: pointerdown, target: Element node <div id=\\"pointerDownTarget\\"></div> }, { type: touchstart, target: Element node <div id=\\"pointerDownTarget\\"></div> }, { type: pointermove, target: Element node <div id=\\"pointerDownTargetParent\\"></div> }, { type: touchmove, target: Element node <div id=\\"pointerDownTarget\\"></div> }, { type: pointerup, target: Element node <div id=\\"pointerDownTargetParent\\"></div> }, { type: touchend, target: Element node <div id=\\"pointerDownTarget\\"></div> } ]" but got "[ { type: pointerdown, target: Element node <div id=\\"pointerDownTarget\\"></div> }, { type: touchstart, target: Element node <div id=\\"pointerDownTarget\\"></div> }, { type: pointermove, target: Element node <div id=\\"pointerDownTargetParent\\"></div> }, { type: pointerup, target: Element node <div id=\\"pointerDownTargetParent\\"></div> }, { type: touchend, target: Element node <div id=\\"pointerDownTarget\\"></div> } ]"
[FAIL] After a touchstart listener moves the target to different position, touchmove event should be fired on the pointerdown target parent
  assert_equals: expected "[ { type: pointerdown, target: Element node <div id=\\"pointerDownTarget\\"></div> }, { type: touchstart, target: Element node <div id=\\"pointerDownTarget\\"></div> }, { type: pointermove, target: Element node <div id=\\"pointerDownTargetParent\\"></div> }, { type: touchmove, target: Element node <div id=\\"pointerDownTarget\\"></div> }, { type: pointerup, target: Element node <div id=\\"pointerDownTargetParent\\"></div> }, { type: touchend, target: Element node <div id=\\"pointerDownTarget\\"></div> } ]" but got "[ { type: pointerdown, target: Element node <div id=\\"pointerDownTarget\\"></div> }, { type: touchstart, target: Element node <div id=\\"pointerDownTarget\\"></div> }, { type: pointermove, target: Element node <div id=\\"pointerDownTargetParent\\"></div> }, { type: pointerup, target: Element node <div id=\\"pointerDownTargetParent\\"></div> }, { type: touchend, target: Element node <div id=\\"pointerDownTarget\\"></div> } ]"
Harness: the test ran to completion.
